general file reorg and cleanup
delete unused pages
move View pages to view directory and rename
pull router/store up to top level
move mixin to utils
fix 'make all'
remove unused package and 'npm preview' command
