+++ src/js/package.json
@@ -6,7 +6,6 @@
   "scripts": {
     "dev": "vite",
     "build": "vite build",
-    "preview": "vite preview",
     "format": "prettier . --write"
   },
   "dependencies": {
@@ -19,7 +18,6 @@
     "@xterm/xterm": "^5.5.0",
     "ace-builds": "^1.41.0",
     "axios": "^1.9.0",
-    "brace": "^0.11.1",
     "buefy": "^1.0.1",
     "bulma": "^0.9.4",
     "d3": "^7.9.0",
