+++ src/js/src/components/MiniTerminal.vue
@@ -4,7 +4,7 @@
 
 <script>
   import axiosInstance from '@/utils/axios.js';
-  import { usePhenixStore } from '@/stores/phenix.js';
+  import { usePhenixStore } from '@/store.js';
 
   import { Terminal } from '@xterm/xterm';
   import { FitAddon } from '@xterm/addon-fit';
