--- src/js/src/components/TestComponent.vue
@@ -1,3 +1,0 @@
-<template>
-  <p>THIS IS A TEST COMPONENT</p>
-</template>
