+++ src/js/src/views/LogsView.vue
